Update query loader Moved queries to a seperate non-server only file to allow for using draft headers on single loaders
