+++ components/Pages/ArchiveRoute.tsx
@@ -1,14 +1,14 @@
 import { notFound } from 'next/navigation'
 
 import { Page } from '@/components/Pages/Page'
-import { loadArchive } from '@/sanity/queries/loadQuery'
+import { loadSingle_Archive } from '@/sanity/queries/loader'
 
 type Props = {
 	params: { slug: string }
 }
 
 export const ArchiveRoute = async ({ params }: Props) => {
-	const data = await loadArchive(params.slug)
+	const data = await loadSingle_Archive(params.slug)
 
 	if (!data) {
 		notFound()
