+++ components/Pages/Blocks.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import dynamic from 'next/dynamic'
-import { loadSettings } from '@/sanity/queries/loadQuery';
 import { BlockList } from './blocks/'
+import { load_Settings } from '@/sanity/queries/loader';
 
 export const Blocks = async ({ blocks, blockClasses }: { blocks: BLOCK_TYPES, blockClasses?: string }) => {
-	const settings = await loadSettings()
+	const settings = await load_Settings()
 
 	return blocks.map((block, i) => {
 		const BlockComponent = BlockList[block._type] ??
