+++ components/Pages/blocks/Archive/Archive.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import { loadArticles } from '@/sanity/queries/loadQuery';
 import { Filter } from '@/components/Pages/blocks/Archive/Filter';
+import { loadBundle_Articles } from '@/sanity/queries/loader';
 
 export const Archive = async ({ data, className }: { data: block_Archive, className?:string }) => {
 	if (!data) return
 
-	const articles = await loadArticles<article>(data.archiveType, data.featuredTaxonomies)
+	const articles = await loadBundle_Articles<article>(data.archiveType, data.featuredTaxonomies)
 	if (!articles) return []
 
 	// const archiveTitle = `${pluralize(camelCaseToWords(data.archiveType))} Archive`
